brains/requirements.txt
Django>=3.2
djangorestframework
django-cors-headers

